Move PathUpdateData to path.h; Cleanup
